requirements.txt
# Basketball Fan Retention & Revenue Optimization - Python 3.10
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
imbalanced-learn>=0.11.0  # for SMOTE
lifelines>=0.27.0  # survival analysis

# Model Interpretation
shap>=0.42.0

# Optimization
pulp>=2.7.0

# Web scraping & API
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ratelimit>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
nbconvert>=7.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
